Move replica related struct to replica.h
