+++ mooncake-store/include/allocation_strategy.h
@@ -6,6 +6,7 @@
 #include <unordered_map>
 
 #include "allocator.h"  // Contains BufferAllocator declaration
+#include "replica.h"
 #include "types.h"
 
 namespace mooncake {
