+++ mooncake-store/include/replica.h
@@ -13,6 +13,76 @@
 
 namespace mooncake {
 
+/**
+ * @brief Type of buffer allocator used in the system
+ */
+enum class ReplicaType {
+    MEMORY,  // Memory replica
+    DISK,    // Disk replica
+};
+
+/**
+ * @brief Stream operator for ReplicaType
+ */
+inline std::ostream& operator<<(std::ostream& os,
+                                const ReplicaType& replicaType) noexcept {
+    static const std::unordered_map<ReplicaType, std::string_view>
+        replica_type_strings{{ReplicaType::MEMORY, "MEMORY"},
+                             {ReplicaType::DISK, "DISK"}};
+
+    os << (replica_type_strings.count(replicaType)
+               ? replica_type_strings.at(replicaType)
+               : "UNKNOWN");
+    return os;
+}
+
+/**
+ * @brief Status of a replica in the system
+ */
+enum class ReplicaStatus {
+    UNDEFINED = 0,  // Uninitialized
+    INITIALIZED,    // Space allocated, waiting for write
+    PROCESSING,     // Write in progress
+    COMPLETE,       // Write complete, replica is available
+    REMOVED,        // Replica has been removed
+    FAILED,         // Failed state (can be used for reassignment)
+};
+
+/**
+ * @brief Stream operator for ReplicaStatus
+ */
+inline std::ostream& operator<<(std::ostream& os,
+                                const ReplicaStatus& status) noexcept {
+    static const std::unordered_map<ReplicaStatus, std::string_view>
+        status_strings{{ReplicaStatus::UNDEFINED, "UNDEFINED"},
+                       {ReplicaStatus::INITIALIZED, "INITIALIZED"},
+                       {ReplicaStatus::PROCESSING, "PROCESSING"},
+                       {ReplicaStatus::COMPLETE, "COMPLETE"},
+                       {ReplicaStatus::REMOVED, "REMOVED"},
+                       {ReplicaStatus::FAILED, "FAILED"}};
+
+    os << (status_strings.count(status) ? status_strings.at(status)
+                                        : "UNKNOWN");
+    return os;
+}
+
+/**
+ * @brief Configuration for replica management
+ */
+struct ReplicateConfig {
+    size_t replica_num{1};
+    bool with_soft_pin{false};
+    std::string preferred_segment{};  // Preferred segment for allocation
+
+    friend std::ostream& operator<<(std::ostream& os,
+                                    const ReplicateConfig& config) noexcept {
+        return os << "ReplicateConfig: { replica_num: " << config.replica_num
+                  << ", with_soft_pin: " << config.with_soft_pin
+                  << ", preferred_segment: " << config.preferred_segment
+                  << " }";
+    }
+};
+
 struct MemoryReplicaData {
     std::vector<std::unique_ptr<AllocatedBuffer>> buffers;
 };
